Работает на больших файлах меньше 1Гб

+++ GZipper/Constants.cs
@@ -8,7 +8,7 @@ namespace GZipper
 {
     public static class Constants
     {
-        public const int BlockLength = 1024 * 1024;
+        public const long BlockLength = 1024 * 1024;
         public const string Compress = "compress";
         public const string Decompress = "decompress";
         public static readonly int ThreadCount = Environment.ProcessorCount;
